+++ src/pages/guard/hello-call-apply/components/index.ts
@@ -1,3 +1,5 @@
 export * from './service-period';
 export * from './service-time';
 export * from './service-total';
+export * from './select-senior';
+export * from './tell-to-sinitto';
