+++ src/pages/guard/hello-call-apply/components/select-senior/index.ts
@@ -1,0 +1,1 @@
+export { SelectSenior } from './SelectSenior';
