+++ src/pages/guard/hello-call-apply/components/tell-to-sinitto/index.ts
@@ -1,0 +1,1 @@
+export { TellToSinitto } from './TellToSinitto';
